--- atd-vze/src/views/Locations/LocationsQA.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-import { withApollo } from "react-apollo";
-
-function LocationsQA(props) {
-  return <div className="animated fadeIn">Not Yet Implemented</div>;
-}
-
-export default withApollo(LocationsQA);
